address Srinivas' comments
